use [[:space:]] instead of \s in grep -E
Older versions of GNU grep don't support \s using ERE (-E flag). In newer versions it is just a substitution for [[:space:]], so use [[:space:]] directly. An alternate fix would be to use grep -P to use Perl regular expressions, but that flag is not available on all platforms (such as BSD and Mac OS X 10.8+)