change zdiff regex to work with older zdiff
Older versions of zdiff (verified on with gzip-1.3.5-13.el5.centos.x86_64 and diffutils-2.8.1-16.el5.x86_64) don't like the tabs in this and start considering POS ID REF to be file names, leading to errors (can be seen in snpcall test). Probably a zdiff bug, but with this change it works in the version from gzip-1.3.5-13 and newer (verified with gzip-1.5-7.el7.x86_64 and diffutils-3.3-4.el7.x86_64).